Rename the library LibreAuth and re-license it to CeCILL.
The "Rust Two-Factor Authentication" (R2FA) name has several drawbacks.
The most obvious one is the use of "Rust" which falsely implies the
library is only compatible with Rust projects. The second one is the use
of "Two-Factor Authentication" which is a restriction of usage: although
it is mainly intended to be used a a collection of second authentication
factors, it may be used as a single authentication method or within a
chain of factors. This restriction of usage also prevents from future
inclusion of different authentication methods. For these reasons, the
name had to be changed. The new one uses the now common "Libre*" pattern
derived from the "Open*" one and the generic "Auth" word. This new name
focus on the free-software aspect and the authentication domain without
limitations.

The change of name is a good occasion for a change of license. The ISC
license is nice but too short and maybe not adapted to local laws. With
the use of the CeCILL license, it remains a free software but takes the
French law is consideration.
http://www.cecill.info/