Replace the rust-crypto crate
The rust-crypto crate does not seems to be maintained anymore.
Because of its monolithic architecture, it is not possible to
expand it any further without forking it. Furthermore, many
important issues are still pending. Because of all of this,
rust-crypto is not a good choice anymore.
In issue #2 ring was suggested as a replacement. A fork using
it already exists. However, ring has two main downsides. First of
all, it deprecates SHA1 which is the default hashing function for
HOTP and TOTP. If SHA1 is to be removed, it would break
compatibility with existing systems. The second downside is, like
rust-crypto, its monolithic architecture which does not enable
third-party addons. Such addons would have enabled the possibility
to add back SHA1 support.
On the opposite, the @RustCrypto project seems to be a good
replacement. Development seems to be active and is has a modular
architecture.