Rewrite the entire password module.
The previous approach of the password module was very limited. For
various reasons, users may want to use a configuration different from
the default ones, which wasn't possible. This rewrite brought the same
interface style that is used on the OATH module, allowing users to tune
the configuration for their needs while providing sane defaults.
While being a functional commit, this code is still a work in progress.