Add the CeCILL-C licensing
The CeCILL 2.1 license isn't suited for a wild use of the library since
it is a "viral" license. Restricting the license of software that links
to it is not the way to go.
The CeCILL-C does not have such restriction, it allows software under
any license to link to the library.
Just in case some people will rely solely on the CeCILL 2.1 license, it
is not removed. Now LibreAuth is dual licensed: either CeCILL-C and
CeCILL 2.1 may be used.