Move the PHC parser to nom5
Although nom 5 has not been released yet, an alpha version is available.
This new version brings an awesome functionality: the use of functions
instead of macros.
Because when a macro uses an other macro, the two must be imported, it
may cause troubles when using different minor versions of nom. For
instance, it already created a compiler warning (unused import) when, in
a newer minor version, a macro ceased to use an other macro.
Using functions instead of macros ins not the only reason to use nom
5.0.0-alpha1: migrating a real parser in a project is a great source of
feedback for nom's author. In this case, it highlighted the need for a
`opt` function behaving the way the `opt!` macro does.

https://twitter.com/gcouprie/status/1118567787172175872
https://twitter.com/gcouprie/status/1118500746125877248